+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Suppressing fields on export - integrate with BE. Refs UIDEXP-375.
 * Reliably show paginator on the 'View all' page. Refs UIDEXP-379.
 * React v19: refactor away from default props for functional components.фвв Refs UIDEXP-376.
+* Add info tip to field suppression. Refs UIDEXP-377
 * Update translation file with item's circulation notes. Refs UIDEXP-294.
 
 ## [6.1.0](https://github.com/folio-org/ui-data-export/tree/v6.1.0) (2024-03-19)